Revert "feat(Go-TCK): Implement `JSON-RPC` methods for `AccountCreateTransaction`, `AccountUpdateTransaction`, `AccountDeleteTransaction` and `AccountInfoQuery` (#879)"
This reverts commit a91e5351f208bcef8dc8106d09ca57495d15255f.
