+++ .gitignore
@@ -18,4 +18,4 @@ vendor/
 .idea
 .env
 go.work
-go.work.sum
+
