--- tck/param/sdk.go
@@ -1,9 +1,0 @@
-package param
-
-type SetupParams struct {
-	OperatorAccountId  string
-	OperatorPrivateKey string
-	NodeIp             *string
-	NodeAccountId      *string
-	MirrorNetworkIp    *string
-}
