--- tck/response/sdk.go
@@ -1,16 +1,0 @@
-package response
-
-type SetupResponse struct {
-	Message string
-	Status  string
-}
-
-func NewSetupReponse(message string) *SetupResponse {
-	response := &SetupResponse{
-		Status: "SUCCESS",
-	}
-	if message != "" {
-		response.Message = message
-	}
-	return response
-}
